[#1268] move udp services to udp_tracker_core package
