+++ src/packages/udp_tracker_core/mod.rs
@@ -1,1 +1,2 @@
+pub mod services;
 pub mod statistics;
