+++ src/servers/udp/mod.rs
@@ -642,7 +642,6 @@ pub mod error;
 pub mod handlers;
 pub mod peer_builder;
 pub mod server;
-pub mod services;
 
 pub const UDP_TRACKER_LOG_TARGET: &str = "UDP TRACKER";
